# server/docs/swagger.yaml
basePath: /
definitions:
...
  response.PageResp:
    properties:
      count:
        description: 总数
        type: integer
      lists:
        description: 数据列表
      pageNo:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页数量
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  schema.SystemCornResp:
    properties:
      cornExpr:
        description: corn表达式
        type: string
      createTime:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人
        type: string
      createdByUser:
        description: 创建人
        type: string
      id:
        type: string
      status:
        description: 状态
        type: number
      taskCode:
        description: 任务编码
        type: string
      taskName:
        description: 任务名称
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
info:
  contact:
    email: x@adtk.cn
    name: xh
    url: http://x.adtk.cn
  description: x_admin后台管理系统
  license:
    name: MIT License
    url: https://gitee.com/xiangheng/x_admin/blob/main/LICENSE
  termsOfService: http://x.adtk.cn
paths:
...
  /api/admin/system_corn/add:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 任务名称
        in: body
        name: TaskName
        schema:
          type: string
      - description: 任务编码
        in: body
        name: TaskCode
        schema:
          type: string
      - description: corn表达式
        in: body
        name: CornExpr
        schema:
          type: string
      - description: 状态
        in: body
        name: Status
        schema:
          type: number
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务新增
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/del:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: taskid
        in: body
        name: Id
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务删除
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/del_batch:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 逗号分割的id
        in: body
        name: Ids
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务删除-批量
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/detail:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: taskid
        in: query
        name: Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/schema.SystemCornResp'
              type: object
      summary: 定时任务详情
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/edit:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: taskid
        in: body
        name: Id
        schema:
          type: string
      - description: 任务名称
        in: body
        name: TaskName
        schema:
          type: string
      - description: 任务编码
        in: body
        name: TaskCode
        schema:
          type: string
      - description: corn表达式
        in: body
        name: CornExpr
        schema:
          type: string
      - description: 状态
        in: body
        name: Status
        schema:
          type: number
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务编辑
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/export_file:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 任务名称
        in: query
        name: TaskName
        type: string
      - description: 任务编码
        in: query
        name: TaskCode
        type: string
      - description: corn表达式
        in: query
        name: CornExpr
        type: string
      - description: 状态
        in: query
        name: Status
        type: number
      - description: 创建人
        in: query
        name: CreatedBy
        type: string
      - description: 创建人名称
        in: query
        name: Nickname
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/octet-stream
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: file
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务导出
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/getTaskList:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取任务列表
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/import_file:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 导入文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 定时任务导入
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/list:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 任务名称
        in: query
        name: TaskName
        type: string
      - description: 任务编码
        in: query
        name: TaskCode
        type: string
      - description: corn表达式
        in: query
        name: CornExpr
        type: string
      - description: 状态
        in: query
        name: Status
        type: number
      - description: 创建人
        in: query
        name: CreatedBy
        type: string
      - description: 创建人名称
        in: query
        name: Nickname
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResp'
                  - properties:
                      lists:
                        items:
                          $ref: '#/definitions/schema.SystemCornResp'
                        type: array
                    type: object
              type: object
      summary: 定时任务列表
      tags:
      - system_corn-定时任务
  /api/admin/system_corn/list_all:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 任务名称
        in: query
        name: TaskName
        type: string
      - description: 任务编码
        in: query
        name: TaskCode
        type: string
      - description: corn表达式
        in: query
        name: CornExpr
        type: string
      - description: 状态
        in: query
        name: Status
        type: number
      - description: 创建人
        in: query
        name: CreatedBy
        type: string
      - description: 创建人名称
        in: query
        name: Nickname
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.SystemCornResp'
                  type: array
              type: object
      summary: 定时任务列表-所有
      tags:
      - system_corn-定时任务
...
swagger: "2.0"